Refac: Naming 수정
