--- src/main/kotlin/com/blife/blife/domain/book/external/booksearchapi/kakao/dto/KakaoBookSearchResponseDocuments.kt
@@ -1,10 +1,0 @@
-package com.blife.blife.domain.book.external.booksearchapi.kakao.dto
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class KakaoBookSearchResponseDocuments(
-	val documents: List<KakaoBookSearchResponse>,
-	val meta: KakaoBookSearchResponseMeta
-)
